Extract /import-config/* routes into blueprints/import_config_routes.py (PR E of quickstart.py refactor)
Moves all four /import-config/* routes plus their immediate helpers into a
dedicated blueprint.  This is the biggest single PR in the refactor sprint
(largest contiguous extraction target identified in the audit).

## What moved

Four routes:
* POST /import-config/preview         -> import_config_preview (the 598-line beast)
* GET  /import-config/report          -> import_config_report
* POST /import-config/preview-mapped  -> import_config_preview_mapped
* POST /import-config/confirm         -> import_config_confirm

Plus their immediate helpers:
* _import_preview_json_default, _coerce_validation_response_payload
* _map_playlist_libraries, count_annotated_lines

## DRY refactors landed alongside the move

The audit flagged this cluster as having significant intra-route
duplication.  Three byte-for-byte-identical inner closures were hoisted
to module scope and deduplicated:

1. `parse_list` (6 lines) -- inner closure in THREE routes (preview,
   preview_mapped, confirm).  Hoisted to module level as
   `_parse_csv_or_list_to_set` (renamed for clarity since 'parse_list'
   is misleading; it parses CSV-or-list-into-set).  18 lines of
   duplication eliminated.

2. `parse_base_plex_libraries` (13 lines) -- inner closure in TWO
   routes (preview, confirm).  Hoisted to module level as
   `_parse_base_plex_libraries`.  26 lines of duplication eliminated.

3. `count_annotated_lines` (14 lines) -- existed at BOTH module level
   in quickstart.py and as an inner closure in import_config_preview,
   byte-for-byte identical.  The inner copy is removed; both call sites
   (preview, preview_mapped) now use the blueprint module-level version.
   14 lines of duplication eliminated.

Total: ~58 lines of literal duplication gone.

## Bug fix bundled in

import_config_confirm had this gem at line 3358:

    if fonts_copied:
        global _FONT_CACHE
        _FONT_CACHE = {}

`_FONT_CACHE` is a module-level dict in modules/assets.py.  The
`global _FONT_CACHE; _FONT_CACHE = {}` in quickstart.py was creating a
NEW module-level global in quickstart.py named _FONT_CACHE that nothing
else ever read -- the real cache lives in modules/assets.py and is
served exclusively via modules.assets.list_overlay_fonts().  Result:
after importing fonts via the config-import flow, the font cache was
NEVER invalidated, and freshly imported fonts wouldn't appear in
overlay dropdowns until the next process restart.

modules/assets.py already has a dedicated `clear_font_cache()`
function for exactly this purpose.  The replacement is one line:

    if fonts_copied:
        assets.clear_font_cache()

Documented inline at the call site with a comment explaining the bug
and why the new version is correct.  No existing test exercises this
behavior, but it's a real bug worth fixing while we're in the area.

## Cross-module dependencies (lazy _qs imports)

Six bundle/overlay-image helpers still live in quickstart.py (PR G
territory).  They're accessed via lazy `import quickstart as _qs`
inside import_config_preview to avoid a load-order cycle:

* _is_allowed_bundle_member
* _is_bundled_overlay_image_archive_member
* _normalize_bundle_member_name
* _yaml_path_suffix
* _rewrite_bundle_library_paths
* _rewrite_bundle_overlay_image_paths

After PR G lands, the lazy-import shim can be replaced with a direct
import from modules.bundle_artifacts (or wherever PR G files them).

## File size note

`blueprints/import_config_routes.py` is ~1,370 lines, well above the
600-line soft-limit.  Documented in the module docstring as a deliberate
cohesion tradeoff:

* The four routes share a tight session-cache flow (import_preview_token,
  import_preview_path, import_preview_*_url/token) and a shared helper
  layer.  Splitting them into separate per-route modules would either
  duplicate the helpers or require a sub-package with relative imports
  -- both worse than keeping them together.
* import_config_preview is 595 lines in a single function -- that one
  mega-function is the real elephant and a future PR can decompose it
  internally if desired.

## Tests

* 791 tests pass; 3 caught the missing-lazy-_qs-imports bug on the
  first run and confirm the fix is correct.
* ruff check clean on both touched files.
* App imports + all 4 routes register correctly in app.url_map.
* AST-verified verbatim extraction for the 5 functions that should not
  have changed.  The 3 that intentionally changed (preview, preview_mapped,
  confirm) have explicitly documented diffs.

## Size impact

* quickstart.py: 8,709 -> 7,397 lines (-1,312, -15.1%)
* blueprints/import_config_routes.py: 1,374 lines (new)
* tests/conftest.py: +16 lines (import_config_routes_module fixture)

## Cumulative refactor progress

Starting baseline (post PR #1320): 11,087 lines in quickstart.py.
After PR A (#1321), PR C (#1322), PR D (#1323), and now PR E:
  quickstart.py: 11,087 -> 7,397 (-3,690 lines, -33.3%)
We're a third of the way to the audit's long-term target.
